+++ trunk/ucon64/src/misc/getopt2.c
@@ -60,7 +60,6 @@ Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
 #endif
 #endif
 #include "misc/file.h"
-#include "misc/getopt.h"                        // struct option
 #include "misc/getopt2.h"
 
 
@@ -446,7 +445,7 @@ getopt2_file_recursion (const char *fname, int (*callback_func) (const char *),
       char buf[FILENAME_MAX], *p;
 
 #if     defined __MSDOS__ || defined _WIN32 || defined __CYGWIN__
-      char c = (char) toupper (path[0]);
+      char c = (char) toupper ((int) path[0]);
       if (path[strlen (path) - 1] == DIR_SEPARATOR ||
           (c >= 'A' && c <= 'Z' && path[1] == ':' && path[2] == 0))
 #else
